+++ pkg/cmd/inventoryFlags.go
@@ -8,7 +8,9 @@ import (
 	"io/ioutil"
 	"mime"
 	"os"
+	"path"
 	"path/filepath"
+	"strings"
 
 	"github.com/pkg/errors"
 	"github.com/reubenmiller/go-c8y-cli/pkg/mapbuilder"
@@ -56,6 +58,76 @@ func addDataFlagWithoutTemplates(cmd *cobra.Command) {
 	cmd.Flags().StringP(FlagDataName, "d", "", "json")
 }
 
+// resolveTemplatePath resolves a template path
+func resolveTemplatePath(name string) (string, error) {
+	return matchFilePath(globalFlagTemplatePath, name, ".jsonnet", "ignore")
+}
+
+func matchFilePath(sourceDir string, pattern string, extension, ignoreDir string) (string, error) {
+	// full path
+	if _, err := os.Stat(pattern); err == nil {
+		return pattern, nil
+	}
+
+	// abort if template path does not exist
+	if _, err := os.Stat(sourceDir); err != nil {
+		return "", fmt.Errorf("Template path does not exist. %s", sourceDir)
+	}
+
+	// path exists under template path
+	fullPath := path.Join(sourceDir, pattern)
+	if _, err := os.Stat(fullPath); err == nil {
+		return fullPath, nil
+	}
+
+	// try to resolve path in nested
+	names, err := resolvePaths(sourceDir, pattern, extension, ignoreDir)
+	if err != nil {
+		return "", err
+	}
+
+	if len(names) == 0 {
+		return "", fmt.Errorf("No matching files found")
+	}
+
+	return names[0], nil
+}
+
+// resolvePaths find matching files within a directory. The filenames ca be filtered by pattern and extension
+func resolvePaths(sourceDir string, pattern string, extension string, ignoreDir string) ([]string, error) {
+	files := make([]string, 0)
+
+	err := filepath.Walk(sourceDir, func(path string, info os.FileInfo, err error) error {
+		if err != nil {
+			fmt.Printf("prevent panic by handling failure accessing a path %q: %v\n", path, err)
+			return err
+		}
+
+		if info.IsDir() && info.Name() == ignoreDir {
+			fmt.Printf("skipping a dir without errors: %+v \n", info.Name())
+			return filepath.SkipDir
+		}
+
+		if extension != "" && !strings.HasSuffix(path, extension) {
+			return nil
+		}
+
+		isMatch := false
+		if pattern != "" {
+			if matched, _ := filepath.Match(pattern, info.Name()); matched {
+				isMatch = true
+			}
+		}
+
+		if isMatch {
+			files = append(files, path)
+		}
+
+		return nil
+	})
+	return files, err
+}
+
 func getDataFlag(cmd *cobra.Command) map[string]interface{} {
 	if value, err := cmd.Flags().GetString(FlagDataName); err == nil {
 		return RemoveCumulocityProperties(MustParseJSON(getContents(value)), true)
@@ -76,6 +148,12 @@ func setDataTemplateFromFlags(cmd *cobra.Command, body *mapbuilder.MapBuilder) e
 	}
 
 	if value, err := cmd.Flags().GetString(FlagDataTemplateName); err == nil {
+
+		if fullFilePath, err := resolveTemplatePath(value); err == nil {
+			Logger.Infof("Template file: %s", fullFilePath)
+			value = fullFilePath
+		}
+
 		contents := getContents(value)
 		body.SetTemplate(contents)
 		if err := body.ApplyTemplate(false); err != nil {